More moving of the unit tests.  Its starting to look like we need to break the libraries into 2 libraries.
