+++ test/hardlink/hardlink.remove.spec.js
@@ -1,12 +1,12 @@
-var jsong = require("../../../index");
+var jsong = require("../../index");
 var Model = jsong.Model;
 var expect = require('chai').expect;
-var Cache = require("../../data/Cache");
-var ReducedCache = require("../../data/ReducedCache");
-var Expected = require("../../data/expected");
-var LocalDataSource = require("../../data/LocalDataSource");
+var Cache = require("../data/Cache");
+var ReducedCache = require("../data/ReducedCache");
+var Expected = require("../data/expected");
+var LocalDataSource = require("../data/LocalDataSource");
 var Rx = require("rx");
-var testRunner = require("../../testRunner");
+var testRunner = require("../testRunner");
 var References = Expected.References;
 var Complex = Expected.Complex;
 var Values = Expected.Values;
