+++ test/hardlink/index.js
